optimizacion del codigo, pruebas realizadas con postman. Agregados mensajes de errores. Control de informacion entrante. Terminadas funcionalidades likes, inboxes y favorites, cambios en detalles de los modelos

+++ api/src/controllers/Comment.js
@@ -4,13 +4,22 @@ const { addNotification } = require('./Inboxes');
 
 const addComment = async(req, res, next) => {
     const { idPost, idUser } = req.params
+    const { message } = req.body
     const obj = {};
 
   try {
-    const createdInPost = await Post.findByPk(idPost, { include: [User] });
+    const createdInPost = await Post.findByPk(idPost, { include: [User, Comment] });
     const createdBy = await User.findByPk(idUser);
+    if(!createdInPost || !createdBy) return res.status(404).send("Parametros invalidos")
 
-    const newComment = await Comment.create(req.body);
+    let exist = createdInPost.dataValues.comments.find(e => e.dataValues.message === message)
+    if(exist) {
+      return res.status(400).send("Rechazado, ya existe este comentario en el posteo")
+    }
+
+    const newComment = await Comment.create({
+      message
+    });
     createdBy.addComment(newComment);
     createdInPost.addComment(newComment);
     
@@ -19,8 +28,8 @@ const addComment = async(req, res, next) => {
     obj.rating = newComment.dataValues.rating;
     obj.user = {
       first_name: createdBy.dataValues.first_name,
-      id: createdBy.dataValues.id,
       last_name: createdBy.dataValues.last_name,
+      id: createdBy.dataValues.id,
     };
     obj.notification = false
     
@@ -35,9 +44,18 @@ const addComment = async(req, res, next) => {
   }
 };
 
-const updateComment = (req, res, next) => {
-  const { idComment } = req.params;
+const updateComment = async (req, res, next) => {
+  const { idComment, idUser } = req.params;
   const { message, rating } = req.body;
+  const comment = await Comment.findByPk(idComment, { include: [User] });
+  const ownUser = await User.findByPk(idUser)
+  
+  if(!comment || !ownUser) return res.status(404).send("Datos no encontrados")
+
+  if(comment.dataValues.user.id !== ownUser.id) {
+    return res.status(400).send("Accion denegada, solo el propietario puede actualizar el comentario")
+  }
+
   return Comment.update(
     { message, rating },
     {
@@ -53,15 +71,27 @@ const updateComment = (req, res, next) => {
     .catch((error) => next(error));
 };
 
-const deleteComment = (req, res, next) => {
-  const { idComment } = req.params;
-  return Comment.destroy({
-    where: {
-      id: idComment,
-    },
-  })
-    .then(() => res.status(200).send("Comment deleted successfully"))
-    .catch((error) => next(error));
+const deleteComment = async(req, res, next) => {
+  const { idComment, idUser } = req.params;
+  try {
+    const comment = await Comment.findByPk(idComment, { include: [User] });
+    const user = await User.findByPk(idUser)
+
+    if(!comment || !user) return res.status(404).send("Datos no encontrados")
+
+    if(comment.dataValues.user.id === user.id || user.isAdmin) {
+      await Comment.destroy({
+        where: {
+          id: idComment
+        }
+      });
+    } else return res.status(400).send("Accion denegada, solo el propietario puede eliminar el comentario")
+
+    res.send("Comment deleted successfully")
+    
+  } catch (error) {
+    next(error)
+  }
 };
 
 module.exports = {
